refactoring

+++ anonymize.py
@@ -81,7 +81,7 @@ class Anonymizer:
         # Grab a batch, shuffled according to the provided seed. Note that
         # i-th image: samples[i][0], i-th mask: samples[i][1]
         samples = Loader.get_batch(self.image_paths, self.batch_size, batch, None)
-        source = samples[0, :, :, :]
+        source = samples[0, :, :, :].astype(int)
         samples.astype(float)
         # Cast samples into torch.FloatTensor for interaction with U-Net
         samples = torch.from_numpy(samples)
